Testing profile for application
Added a test profile for the app which uses a postgres test database
with pgvector. For the test profile to be able to run the
ApplicationContext needs to recive the Env Variables DB_DRIVER,
DB_PASSWORD, DB_URL_TEST, DB_USER so that Postgres JDBC reaches and
connects to a test database. Also a schema called embeddings and a
tables that represents the Law model entity is required.