Implement LawDao
Implemented the Law Data Access Object while adhearing to interfaces.
Special method to access Law entites over law code was created for
situation without id access.
In LawDaoIntegrationTest changed the injection to interfaces of the
LawDao insted of injecting the concrete implementation in order to
adhere to dependency inversion principle.